Fixed inputColSize
For horizontal forms inputColSize is an optional integer value used to
determine the width of the input and label columns. Specifiying
inputColSize:9 will produce and input column set to col-sm-9 and a label
column of col-sm-3, where 'col-sm-*' are Twitter Bootstrap grid classes.
The total of the two integers, in this case 3 and 9, must equal 12. This
now produces the correct results.